# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "REST APIs to CASE_EXPED table in SQL",
    "version": "1.0.0",
    "title": "API to CASE_EXPED "
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/GetRequestsByStatusId": {
      "get": {
        "tags": [
          "request"
        ],
        "summary": "Get unassigned Requests ",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "request_status_type_id",
            "type": "string",
            "description": "Get the Request by status Id",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/GetRequests": {
      "get": {
        "tags": [
          "request"
        ],
        "summary": "Get Requests by User id",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "assigned_user_id",
            "type": "string",
            "description": "Get the requests by user id",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/api/GetRequestDetails": {
      "get": {
        "tags": [
          "request"
        ],
        "summary": "Get Requests by ID",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "description": "Get request by id",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/NewRequest": {
      "post": {
        "tags": [
          "request"
        ],
        "summary": "Add a request",
        "description": "",
        "operationId": "addRequest",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request object needed",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Request"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/UpdateRequest": {
      "put": {
        "tags": [
          "request"
        ],
        "summary": "Update a request",
        "description": "",
        "operationId": "updateRequest",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request object needed",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Request"
            }
          },
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "description": "Update request by id",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/api/GetSolutionByID": {
      "get": {
        "tags": [
          "solution"
        ],
        "summary": "Get Solution by ID",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "description": "Get solution by id",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/GetSolutionForRequestID": {
      "get": {
        "tags": [
          "solution"
        ],
        "summary": "Get Solution for Request ID",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "request_id",
            "type": "string",
            "description": "Get solution for request id",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/AddSolution": {
      "post": {
        "tags": [
          "solution"
        ],
        "summary": "Add a solution",
        "description": "",
        "operationId": "addSolution",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request object needed",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Solution"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/api/AddCarrier": {
      "post": {
        "tags": [
          "carrier"
        ],
        "summary": "Add new carrier",
        "description": "",
        "operationId": "addCarrier",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request object needed",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Carrier"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/api/AddCSAT": {
      "post": {
        "tags": [
          "csat"
        ],
        "summary": "Add CSAT",
        "description": "",
        "operationId": "addCSAT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Request object needed",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CSAT"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Request": {
      "type": "object",
      "properties": {
        "customer_id": {
          "type": "integer",
          "format": "int64"
        },
        "from_address_city": {
          "type": "string",
          "description": "Origin City Name "
        },
        "from_address_country": {
          "type": "string",
          "description": "Origin City Country "
        },
        "from_address_postcode": {
          "type": "string",
          "format": "Origin postcode"
        },
        "to_address_city": {
          "type": "string",
          "description": "Destination City Name "
        },
        "to_address_country": {
          "type": "string",
          "description": "Origin City Country "
        },
        "to_address_postcode": {
          "type": "string",
          "format": "Destination postcode"
        },
        "load_datetime": {
          "type": "string",
          "description": "Loading datetime"
        },
        "unload_datetime": {
          "type": "string",
          "format": "Unloading datetime"
        },
        "solution_time": {
          "type": "string",
          "description": "Solution Time"
        },
        "validation_datetime": {
          "type": "string",
          "format": "Validaiton datetime"
        },
        "goods_weight": {
          "type": "string",
          "description": "Goods Weight"
        },
        "goods_europallets": {
          "type": "string",
          "format": "Googs/europallets"
        },
        "truck_type_id": {
          "type": "string",
          "description": "Truck Type Id"
        },
        "adr_type_id": {
          "type": "string",
          "format": "Address type"
        },
        "special_requirments": {
          "type": "string",
          "description": "Special Requirments"
        },
        "email_html": {
          "type": "string",
          "format": "Googs/europallets"
        },
        "assigned_user_id": {
          "type": "string",
          "description": "Truck Type Id"
        },
        "request_source_id": {
          "type": "string",
          "description": "Request Source"
        },
        "transporeon_id": {
          "type": "string",
          "description": "Truck Type Id"
        },
        "request_status_type_id": {
          "type": "string",
          "description": "Order Status"
        }
      },
      "xml": {
        "name": "Request"
      }
    },
    "Solution": {
      "type": "object",
      "properties": {
        "request_id": {
          "type": "integer",
          "format": "int64"
        },
        "price": {
          "type": "string"
        },
        "delay": {
          "type": "string"
        },
        "carrier_id": {
          "type": "string"
        },
        "truck_type_id": {
          "type": "string"
        },
        "details": {
          "type": "string"
        },
        "transit_time": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Solution"
      }
    },
...
    "Carrier": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "orders_fulfilled": {
          "type": "string"
        },
        "rating": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tracking"
      }
    },
...
    "CSAT": {
      "type": "object",
      "properties": {
        "customer_id": {
          "type": "integer",
          "format": "int64"
        },
        "score": {
          "type": "string"
        },
        "request_id": {
          "type": "integer",
          "format": "int64"
        },
        "feedback": {
          "type": "string"
        }
      },
      "xml": {
        "name": "CSAT"
      }
    }
  },
...
}